:whale: Ensure uwsgi becomes PID 1
Before this patch, uwsgi was not running as PID, but the docker_start
script got PID 1. This means that when the container runtime signals
to terminate the container, the signal was not actually passed down.

Container runtimes then usually follow that up with a SIGKILL (after
the SIGTERM), after a grace period.

Container restarts should be near-instant after this fix, because the
SIGTERM will be honored instead of having to wait the grace period and
have the container killed by SIGKILL.