:bug: Fix resource leak bug
When a result backend is stored, celery tracks task execution
meta-data and the result of the task. This should either be fetched
and cleared from the store, or not stored at all - which is what we
opt for here.

Redis *should* evict keys by itself, but in combination with the
previous RabbitMQ result backend (see bde607667ec41dd2218a05d49584716668e2ed08),
this caused excessive resource consumption and actual shutdowns of
long-running RabbitMQ instances.